Fix: parent height on template for an edge case
When a child (among 2 within a flex container) has vertically centered content, its top content is cropped if a parent div sets a height to 100% of viewport instead of min-height.
It's a weird behavior I can't explain, because when changing the vertical alignment from centered to top, the crop disappears (although it's not supposed to make any difference).