fix: fix a problem with user metadata state during check session,
fix different bugs related to stripe and how it check if user is Free from the server. those bug happened after the way we check for session changed. most bugs should be resolved, the workflow is working but there might be improvement to this feature to do in the future.